tools/grammar_tool: fix set crossproduct
This special case where lhs is empty originated when trying to .fold() over a
list of tokens. The initial value of the fold was an empty set, when it should
have been a set containing the empty string. That fix is in the next commit and
I'm too lazy to refactor/rebase it into this one.