phase3: fix lexing of /* old style comments */
The `^` which roots a regex to the beginning of the string does not distribute
over the `|` alternation. So it would only match `//` comments that were at the
beginning of the string but would match `/**/` comments anywhere in the string

Apply the same fix (using parenthesis) to the regex for identifiers as well in
case we do ever get around to adding `|` alternation for unicode characters? Not
sure how that would work anyway.