Rework token tracking
This is extremely cathartic. The past 10 commits have been building up
to this. Some of those were distractions, and others felt like necessary
prerequisites. Ultimately this has been the goal for quite a while
though. The aim has been to improve error reporting.

the old Location was the most important thing to get rid of. It
conflated several different things. Now we have TokenOrigin, which
specifies whether a token appeared in this exact location in the input
file or whether it is the result of a macro expansion. MacroResult +
MacroInvocation will be able to track the flow of a token through macros.